requirements.txt
# Core dependencies
tensorflow
numpy==1.23.5
Pillow==9.3.0
matplotlib==3.6.2
scikit-learn==1.1.3
pandas==1.5.2
opencv-python==4.6.0.66

# For the web application
streamlit==1.15.2
watchdog==2.1.9

# Development tools
pytest==7.2.0
black==22.10.0
flake8==6.0.0

# Utilities
tqdm==4.64.1
h5py==3.7.0
pyyaml==6.0

# Other dependencies
requests==2.28.1
beautifulsoup4==4.11.1
dotenv==1.0.1
